\name{optimgui}
\alias{optimgui}
\title{A GUI to assist optimization problems in R}
\usage{
optimgui()
}
\description{
This function will invoke the main GUI window of optimgui program.
}
\details{
This package aims at building up a GUI to assist the preparation and solution of optimization problems, and optimgui() is the only exported function in this package.

See the package vignette for detailed usages.
}
\author{
Yixuan Qiu \email{yixuan.qiu [at] cos.name}

John C. Nash \email{nashjc [at] uottawa.ca}
}
\keyword{optimize}
\seealso{
\code{\link[optimx]{optimx}}, \code{\link[stats]{constrOptim}}, \code{\link[Rsolnp]{solnp}}
}
\examples{
\dontrun{optimgui()}
}
